DonorsApp/requirements.txt
Flask==2.0.3
# Flask-MySQLdb==1.0.1
mysql-connector-python==8.0.28
pyaml
